Remove repeated styles
