+++ src/components/EditCoursePage/EditCoursePage.js
@@ -1,36 +1,6 @@
 import React from "react";
-import { withStyles } from "@material-ui/core/styles";
 import { withState } from "../../utils/State";
 import CourseForm from "../CourseForm/CourseForm";
-
-const drawerWidth = 240;
-
-const styles = theme => ({
-  drawerHeader: {
-    display: "flex",
-    alignItems: "center",
-    padding: theme.spacing(0, 1),
-    ...theme.mixins.toolbar,
-    justifyContent: "flex-end",
-  },
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing(3),
-    transition: theme.transitions.create("margin", {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-    marginLeft: 0,
-  },
-  contentShift: {
-    transition: theme.transitions.create("margin", {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-    marginLeft: drawerWidth,
-  },
-});
-
 class EditCoursePage extends React.Component {
   constructor(props) {
     super(props);
@@ -38,16 +8,13 @@ class EditCoursePage extends React.Component {
   }
 
   render() {
-    const { classes, history, context } = this.props;
+    const { history, context } = this.props;
     return (
       <div>
-        <main className={`${classes.content} ${this.state.open ? classes.contentShift : ""}`}>
-          <div className={classes.drawerHeader} />
-          <CourseForm course={context.course} history={history} />
-        </main>
+        <CourseForm course={context.course} history={history} />
       </div>
     );
   }
 }
 
-export default withState(withStyles(styles)(EditCoursePage));
+export default withState(EditCoursePage);
