+++ src/components/SolveActivityPage/SolveActivityPage.js
@@ -20,30 +20,7 @@ import { prepareInitialCode } from "../../utils/files";
 // Styles
 import "react-mde/lib/styles/css/react-mde-all.css";
 
-const drawerWidth = 240;
-
 const styles = theme => ({
-  drawerHeader: {
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "flex-end",
-    height: 56,
-  },
-  content: {
-    flexGrow: 1,
-    transition: theme.transitions.create("margin", {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-    marginLeft: 0,
-  },
-  contentShift: {
-    transition: theme.transitions.create("margin", {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-    marginLeft: drawerWidth,
-  },
   circularProgress: {
     position: "absolute",
     left: "50%",
@@ -67,7 +44,6 @@ type Props = {
 
 type State = {
   error: { open: boolean, message: ?string },
-  isSideBarOpen: boolean,
   activity: ?Activity,
   code: { [string]: string },
   editorWidth: string,
@@ -81,7 +57,6 @@ type State = {
 class SolveActivityPage extends React.Component<Props, State> {
   state = {
     error: { open: false, message: null },
-    isSideBarOpen: false,
     editorWidth: "100%",
     activity: null,
     code: { "main.c": "" },
@@ -184,7 +159,6 @@ class SolveActivityPage extends React.Component<Props, State> {
     const { classes, history } = this.props;
     const {
       activity,
-      isSideBarOpen,
       submittedActivity,
       selectedSubmissionId,
       editorWidth,
@@ -227,8 +201,7 @@ class SolveActivityPage extends React.Component<Props, State> {
 
         {!activity && <CircularProgress className={classes.circularProgress} />}
         {activity && (
-          <main className={classes.content}>
-            <div className={classes.drawerHeader} />
+          <div>
             <SolvePageHeader
               handleSubmitActivity={e => this.handleSubmitActivity(e)}
               handleOpenPastSubmissionsSidePanel={() => this.setOpenSubmissionsPanel()}
@@ -283,7 +256,7 @@ class SolveActivityPage extends React.Component<Props, State> {
                 <MarkdownRenderer content={activity.description} />
               </div>
             </SplitPane>
-          </main>
+          </div>
         )}
         {submittedActivity && (
           <SubmissionResultModal
