+++ src/components/UsersPage/UsersPage.js
@@ -16,32 +16,7 @@ import ErrorNotification from "../../utils/ErrorNotification";
 
 import type { Student } from "../../types";
 
-const drawerWidth = 240;
-
 const styles = theme => ({
-  drawerHeader: {
-    display: "flex",
-    alignItems: "center",
-    padding: theme.spacing(0, 1),
-    ...theme.mixins.toolbar,
-    justifyContent: "flex-end",
-  },
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing(3),
-    transition: theme.transitions.create("margin", {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-    marginLeft: 0,
-  },
-  contentShift: {
-    transition: theme.transitions.create("margin", {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-    marginLeft: drawerWidth,
-  },
   title: {
     marginTop: 20,
     marginBottom: 20,
@@ -85,14 +60,12 @@ type Props = {
 
 type State = {
   error: { open: boolean, message: ?string },
-  isSideBarOpen: boolean,
   users: Array<Student>,
 };
 
 class UsersPage extends React.Component<Props, State> {
   state = {
     error: { open: false, message: null },
-    isSideBarOpen: false,
     users: [],
   };
 
@@ -175,17 +148,12 @@ class UsersPage extends React.Component<Props, State> {
   render() {
     const { classes } = this.props;
 
-    const { users, isSideBarOpen, error } = this.state;
+    const { users, error } = this.state;
 
     return (
       <div>
         {error.open && <ErrorNotification open={error.open} message={error.message} />}
-        <main className={`${classes.content} ${isSideBarOpen ? classes.contentShift : ""}`}>
-          <div className={classes.drawerHeader} />
-          <div className={classes.tableContainerDiv}>
-            {users && this.renderUsers(users, classes)}
-          </div>
-        </main>
+        <div className={classes.tableContainerDiv}>{users && this.renderUsers(users, classes)}</div>
       </div>
     );
   }
